requirements.txt
fastapi
uvicorn
httpx
python-dotenv

#pip install -r requirements.txt
